requirements.txt
Flask==3.1.2
Jinja2==3.1.6
nltk>=3.8.0
scikit-learn>=1.2.0
joblib>=1.2.0
pdfminer.six>=20221105
google-genai>=0.3.0
requests>=2.28.0
